+++ package.json
@@ -9,6 +9,7 @@
     "@testing-library/user-event": "^14.4.3",
     "localforage": "^1.10.0",
     "match-sorter": "^6.3.1",
+    "prop-types": "^15.8.1",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-redux": "^8.0.5",
