+++ src/features/RocketList/RocketList.css
@@ -1,0 +1,3 @@
+.RocketList {
+  padding: 0 50px 40px;
+}
